# api_docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KrishiSetu AI Backend",
    "description": "Smart India Hackathon 2026 (SIH 26033) - Feature Phone First Farmer Marketplace",
    "version": "1.0.0"
  },
  "paths": {
...
    "/transaction/confirm": {
      "post": {
        "tags": [
          "Transactions & Settlements"
        ],
        "summary": "Confirm Transaction",
        "operationId": "confirm_transaction_transaction_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transaction/verify-otp": {
      "post": {
        "tags": [
          "Transactions & Settlements"
        ],
        "summary": "Verify Otp",
        "operationId": "verify_otp_transaction_verify_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OTPVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sms/send": {
      "post": {
        "tags": [
          "SMS Gateway"
        ],
        "summary": "Send Sms",
        "operationId": "send_sms_sms_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SMSSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sms/reply": {
      "post": {
        "tags": [
          "SMS Gateway"
        ],
        "summary": "Reply Sms",
        "description": "Simulates incoming SMS from feature phone farmer (e.g. 'YES', '1', 'NO').",
        "operationId": "reply_sms_sms_reply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SMSReplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sms/logs": {
      "get": {
        "tags": [
          "SMS Gateway"
        ],
        "summary": "Get Sms Logs",
        "operationId": "get_sms_logs_sms_logs_get",
        "parameters": [
          {
            "name": "phone",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "+919812345001",
              "title": "Phone"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SMSLogResponse"
                  },
                  "title": "Response Get Sms Logs Sms Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OTPVerifyRequest": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "otp_code": {
            "type": "string",
            "title": "Otp Code"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "otp_code"
        ],
        "title": "OTPVerifyRequest"
      },
      "OTPVerifyResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "transaction_status": {
            "type": "string",
            "title": "Transaction Status"
          },
          "escrow_status": {
            "type": "string",
            "title": "Escrow Status"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "transaction_status",
          "escrow_status"
        ],
        "title": "OTPVerifyResponse"
      },
...
      "SMSLogResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "phone",
          "direction",
          "message",
          "status",
          "timestamp"
        ],
        "title": "SMSLogResponse"
      },
      "SMSReplyRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone",
            "default": "+919812345001"
          },
          "reply_text": {
            "type": "string",
            "title": "Reply Text"
          }
        },
        "type": "object",
        "required": [
          "reply_text"
        ],
        "title": "SMSReplyRequest"
      },
      "SMSSendRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "message"
        ],
        "title": "SMSSendRequest"
      },
      "TransactionConfirmRequest": {
        "properties": {
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "buyer_id": {
            "type": "string",
            "title": "Buyer Id"
          },
          "truck_route_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Truck Route Id",
            "default": "TRUCK-001"
          },
          "agreed_price_per_kg": {
            "type": "number",
            "title": "Agreed Price Per Kg"
          },
          "quantity_kg": {
            "type": "number",
            "title": "Quantity Kg"
          }
        },
        "type": "object",
        "required": [
          "listing_id",
          "buyer_id",
          "agreed_price_per_kg",
          "quantity_kg"
        ],
        "title": "TransactionConfirmRequest"
      },
      "TransactionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "listing_id": {
            "type": "string",
            "title": "Listing Id"
          },
          "buyer_id": {
            "type": "string",
            "title": "Buyer Id"
          },
          "farmer_id": {
            "type": "string",
            "title": "Farmer Id"
          },
          "truck_route_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Truck Route Id"
          },
          "quantity_kg": {
            "type": "number",
            "title": "Quantity Kg"
          },
          "agreed_price_per_kg": {
            "type": "number",
            "title": "Agreed Price Per Kg"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "logistics_shared_cost": {
            "type": "number",
            "title": "Logistics Shared Cost"
          },
          "logistics_solo_cost": {
            "type": "number",
            "title": "Logistics Solo Cost"
          },
          "farmer_savings": {
            "type": "number",
            "title": "Farmer Savings"
          },
          "escrow_status": {
            "type": "string",
            "title": "Escrow Status"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "otp_code": {
            "type": "string",
            "title": "Otp Code"
          },
          "otp_verified": {
            "type": "boolean",
            "title": "Otp Verified"
          },
          "pickup_time": {
            "type": "string",
            "title": "Pickup Time"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "listing_id",
          "buyer_id",
          "farmer_id",
          "quantity_kg",
          "agreed_price_per_kg",
          "total_amount",
          "logistics_shared_cost",
          "logistics_solo_cost",
          "farmer_savings",
          "escrow_status",
          "status",
          "otp_code",
          "otp_verified",
          "pickup_time",
          "created_at"
        ],
        "title": "TransactionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}